move export Pdf function to a separate module
